Move get_public_ip to an 'util' file
